♻️ splits clinical typedef to a new file because it was getting too big
